+++ CHANGES.md
@@ -15,7 +15,7 @@
 * Remove the `tracers` router config in favor of the `io.l5d.zipkin` telemeter.
 * Change the default `dstPrefix` from the protocol name to `/svc`
 * Change the default HTTP identifier to the `io.l5d.header.token` identifier
-* Add `io.l5d.promethus` telemeter
+* Add `io.l5d.prometheus` telemeter
 * Remove `io.l5d.commonMetrics` telemeter
 * Remove `io.l5d.adminMetricsExport` as a telemeter plugin, it is now always
   loaded